AwsCryptographicMaterialProviders/runtimes/python/pyproject.toml
[tool.poetry]
name = "aws-cryptographic-material-providers"
version = "0.1.0"
description = ""
authors = ["AWS Crypto Tools <aws-crypto-tools@amazon.com>"]
packages = [
    { include = "aws_cryptographic_materialproviders", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11.0"
# TODO: Depend on PyPi once Smithy-Python publishes their Python package there
smithy-python = { path = "../../../submodules/smithy-python/python-packages/smithy-python", develop = false}
DafnyRuntimePython = "^4.4.0"
standard-library = { path = "../../../StandardLibrary/runtimes/python", develop = false}
amazon-cryptography-internal-dafny-shim-kms = {path = "../../../ComAmazonawsKms/runtimes/python", develop = false}
amazon-cryptography-internal-dafny-shim-ddb = {path = "../../../ComAmazonawsDynamodb/runtimes/python", develop = false}
aws-cryptography-primitives = {path = "../../../AwsCryptographyPrimitives/runtimes/python", develop = false}

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
